requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-redis==0.2.1
langchain-text-splitters==0.3.8
langchain-experimental==0.3.0
langchain-milvus==0.2.0
langgraph==0.2.20
python-multipart==0.0.9
pydantic-settings==2.6.1
python-dotenv==1.0.1
pymilvus==2.6.1
pypdf==5.1.0
pymupdf==1.24.14
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy>=1.24.3
httpx==0.28.1
aiofiles==24.1.0
loguru==0.7.2
tenacity>=9.1.2,<10.0.0
orjson==3.10.11
requests==2.32.3
